+++ cmd/ggg/main.go
@@ -1,10 +1,12 @@
 package main
 
 import (
+	"errors"
 	"fmt"
 	"os"
 
 	"github.com/go-go-golems/infra-tooling/internal/cli"
+	"github.com/go-go-golems/infra-tooling/internal/exitcode"
 )
 
 func main() {
@@ -14,6 +16,10 @@ func main() {
 		os.Exit(1)
 	}
 	if err := root.Execute(); err != nil {
+		var exitErr exitcode.Error
+		if errors.As(err, &exitErr) {
+			os.Exit(exitErr.Code)
+		}
 		os.Exit(1)
 	}
 }